Re-introduced templates to Quaternion functions.
Without templates, unoptimized functions were selected for types
derived from the ones on which the function was defined. All
functions except multiplication are now templated, and tests have
been added to ensure the optimized functions are run, not the
general case. Unfortunately, the current methodology causes an
ambiguous resolution for multiplication, so it has not been
updated yet.